# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Money guard app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of Money Guard App"
  },
  "servers": [
    {
      "url": "https://server-money-guard-teamproject.onrender.com"
    }
  ],
...
  "paths": {
...
    "/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get list of transactions",
        "operationId": "getTransactions",
        "description": "Get list of Transactions with this endpoint",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "description": "All transactions from date"
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "enum": [
                "Main expenses",
                "Products",
                "Car",
                "Self care",
                "Child care",
                "Household products",
                "Education",
                "Leisure",
                "Other expenses",
                "Entertaiment"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found transactions!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found transactions!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/transaction"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Create Transaction",
        "operationId": "createTransaction",
        "description": "Create a transaction with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "date",
                  "type",
                  "sum",
                  "category"
                ],
                "properties": {
                  "date": {
                    "description": "transaction's date",
                    "type": "string",
                    "example": "dd-mm-yyyy"
                  },
                  "type": {
                    "description": "transaction's type",
                    "type": "string",
                    "enum": [
                      "Income",
                      "Expense"
                    ]
                  },
                  "category": {
                    "description": "transaction's category",
                    "type": "string",
                    "enum": [
                      "Main expenses",
                      "Products",
                      "Car",
                      "Self care",
                      "Child care",
                      "Household products",
                      "Education",
                      "Leisure",
                      "Other expenses",
                      "Entertainment"
                    ]
                  },
                  "comment": {
                    "description": "transaction's comment",
                    "type": "string",
                    "example": "payment for first semester"
                  },
                  "sum": {
                    "description": "transaction's sum",
                    "type": "number",
                    "example": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction Successfully Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a transaction!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/transaction"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
...
      "404": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Route not found"
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "transaction": {
        "type": "object",
        "required": [
          "_id",
          "date",
          "type",
          "sum",
          "userId"
        ],
        "properties": {
          "_id": {
            "description": "transaction's id",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "date": {
            "description": "transaction's date",
            "type": "string",
            "example": "22-12-2025"
          },
          "type": {
            "description": "transaction's type",
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "category": {
            "description": "transaction's category",
            "type": "string",
            "enum": [
              "Main expenses",
              "Products",
              "Car",
              "Self care",
              "Child care",
              "Household products",
              "Education",
              "Leisure",
              "Other expenses",
              "Entertaiment",
              "Income"
            ]
          },
          "comment": {
            "description": "transaction's comment",
            "type": "string",
            "example": "payment for first semester"
          },
          "sum": {
            "description": "transaction's sum",
            "type": "number",
            "example": 1000
          },
          "userId": {
            "description": "transaction's user id",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          }
        }
      },
...
    }
  }
}